Truncate referenceDateTime regardless of source
Java 9 introduced time resolution to microseconds. This led to a failing test, since the system properties now contained microseconds, which the production code could not handle. In order to capture the referenceDateTime in a single long in the InitializationCreatorSnapshot, truncating it first is necessary.
The referenceDateTime is now truncated regardless of its source ('now()' or system property). It is especially required for the system property, as this is specified by the user and thus might have any resolution that the format for parsing allows.

Issue: #1
Signed-off-by: Stefan Hechtl <stefan.hechtl@tngtech.com>